+++ server/access.go
@@ -271,6 +271,7 @@ func (accesses *Accesses) Write(db *Database) error {
 	}
 
 	if err != nil {
+		rows.Close()
 		return formatError(err)
 	}