+++ server/apikey.go
@@ -240,6 +240,7 @@ func (apikeys *Apikeys) Write(db *Database) error {
 	}
 
 	if err != nil {
+		rows.Close()
 		return formatError(err)
 	}