+++ server/dirwatch.go
@@ -662,6 +662,7 @@ func (dirwatches *Dirwatches) Write(db *Database) error {
 	}
 
 	if err != nil {
+		rows.Close()
 		return formatError(err)
 	}