+++ server/downstream.go
@@ -474,6 +474,7 @@ func (downstreams *Downstreams) Write(db *Database) error {
 	}
 
 	if err != nil {
+		rows.Close()
 		return formatError(err)
 	}