+++ server/tags.go
@@ -228,6 +228,7 @@ func (tags *Tags) Write(db *Database) error {
 	}
 
 	if err != nil {
+		rows.Close()
 		return formatError(err)
 	}